Changing linked_list to be more dyanmic
Dynamic data size, types, and addition of 'data names' to describe each
field in a node. Load and save no longer work, but add, edit, and delete
do.

The memory address of lists is now limited in scope to linked_list.c,
newlist() returns a reference number to the list that other libraries
can use to retrieve list and node data.